+++ src/main/java/com/becareful/becarefulserver/domain/matching/controller/CompletedMatchingController.java
@@ -1,5 +1,7 @@
 package com.becareful.becarefulserver.domain.matching.controller;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,11 +15,13 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequestMapping("/complete-matching")
 @RequiredArgsConstructor
+@Tag(name = "Complete Matching", description = "확정된 매칭 관련 API 입니다.")
 public class CompletedMatchingController {
 
     private final CompletedMatchingService completedMatchingService;
 
     @PostMapping
+    @Operation(summary = "계약서를 기반으로 매칭을 확정합니다.")
     public ResponseEntity<Void> createCompleteMatching(@RequestParam(name = "contractId") Long contractId) {
         completedMatchingService.createCompletedMatching(contractId);
         return ResponseEntity.ok().build();
